requirements.txt
glob3
numpy
opencv-python
Pillow